% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTimings.R
\name{addToTimingsTable}
\alias{addToTimingsTable}
\title{Time steps in a script}
\usage{
addToTimingsTable(entryName = "Start", timingsTable = NULL, printMsg = F,
  printTime = F)
}
\arguments{
\item{entryName}{The Step to be run}

\item{timingsTable}{A timingsTable to be added to}

\item{printMsg}{Whether or not the \code{entryName} should be printed to the screen}
}
\value{
A data frame containing the tasks performed and when they were started. If \code{timingsTable}
is provided then this Will be appended to.
}
\description{
This function allows you to track the time it takes to run different steps in a script / function / etc.
}
\examples{
# Standard Usage
timingsTable = addToTimingsTable("Start")

# print out the msg
timingsTable = addToTimingsTable("Start", NULL, T)

# Multiple steps
timingsTable = addToTimingsTable("Start")
timingsTable = addToTimingsTable("Another Step", timingsTable)

}

