% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{appendToTable}
\alias{appendToTable}
\title{Append a data set to another data set}
\usage{
appendToTable(full, tmp)
}
\arguments{
\item{full}{A \code{data.frame} that will be the base}

\item{tmp}{A \code{data.frame} that will be joined onto the bottom of \code{full}}
}
\description{
This function takes two data sets and appends the second onto the first.
The Second may not have all of the columns as the firstin a list and converts it to a \code{data.frame}.
}
\examples{
x = data.frame(A = rnorm(3), B = rnorm(3), C = rnorm(3))
y = data.frame(A = rnorm(3), C = rnorm(3))
appendToTable(x, y)

}

