% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{as.data.frame.list}
\alias{as.data.frame.list}
\title{Convert a list to a data frame}
\usage{
\method{as.data.frame}{list}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{list} where all of the elements represents a row and are
typically are of exqual length}
}
\description{
This function takes in a list and converts it to a \code{data.frame}.
}
\examples{
as.data.frame.list(list(x = rnorm(100), y = rnorm(100)))

}

