% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClassification.R
\name{binaryClassifierEffect}
\alias{binaryClassifierEffect}
\title{Binary Classifier Effects}
\usage{
binaryClassifierEffect(probs, df, var.names, max.n = 1000, display = FALSE,
  log.names = NULL)
}
\arguments{
\item{probs}{A vector of estimated probabilities, from 0 to 1.}

\item{df}{A data frame containing variables that were used to predict some binary outcome.}

\item{var.names}{A vector of column names from df.}

\item{max.n}{The maximum number of points to include in the resulting scatter plots.}

\item{display}{A logical, that determines is the resulting plots are displayed or just returned.}

\item{log.names}{A vector of column names to plot in log scale}
}
\value{
A list of ggplot objects, suitable for printing or saving.
}
\description{
Produce a list of ggplots, showing the relationship between each predictor variable and a probability.

Given a data frame and a vector of probabilities, such as those produced by a binary classifier
model, this function produces a list of ggplots, showing the relationship between each of the named
variables and the probability.

It produces a violin plot for factors, a scatter plot with a loess fitted curve for numeric, and an
error otherwise.
}
\examples{
df<-mtcars
df$am<-as.factor(df$am)
model<-glm(vs ~ am*mpg, data=df, family=binomial)
probs<-model$fit
binaryClassifierEffect(probs,df,c('am','mpg'))

}
\seealso{
\code{\link{qplot}}
}

