% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClassification.R
\name{binaryClassifierEmailReport}
\alias{binaryClassifierEmailReport}
\title{Email a Binary Classifier Report}
\usage{
binaryClassifierEmailReport(evaluationResults,
  toAddress = "your.name@gmail.com", additionalMessage = "", plotSize = 10,
  emailFrontEnd = "")
}
\arguments{
\item{evaluationResults}{The result from running \code{binaryClassifierEvaluation}.}

\item{toAddress}{Character vector, containing the email address(es) to send the report to, when completed.}

\item{additionalMessage}{A string that will appear in the subject of the email}

\item{plotSize}{Numeric scalar containing the side length of the resulting square png files, in pixels.}

\item{emailFrontEnd}{A string of html that goes at the front of the email}
}
\value{
The result of the call to \code{sendmail}
}
\description{
Given the result of \code{binaryClassifierEvaluation} email a standardised report to one or more recipients.

The emailed report contains a histogram of estimated probabilities,
the ROC plot with the calculated area under the curve,
two density plots, one violin, one standard,
two loss curves as a function of deciles of the calculated probabilies,
and a plot of cumulative % qualified versus cumulative % targetted.
It contains some stats such as Area under the ROC curve, Root Mean Square Error and the confusion matrix.
It does not require the actual model.  If the model is supplied, a summary
of the model is also included in the output.
}
\examples{
set.seed(123)
probs <- pnorm(rnorm(100,-1,1))
hist(probs)
actuals <- rbinom(100,1,probs)
plot(as.factor(actuals),probs)

binaryClassifierEmailReport(binaryClassifierEvaluation(probs, actuals),"ivan@xxx.com")

}
\seealso{
\code{\link{RQuantSendMail}}, \code{\link{binaryClassifierEvaluation}}
}

