% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEmail.R
\name{buildhtmlmsg}
\alias{buildhtmlmsg}
\title{Build HTML email message}
\usage{
buildhtmlmsg(body, attachmentFileNames = NULL)
}
\arguments{
\item{body}{A \code{list} or \code{vector} of html elements to go into the
main message of the email}

\item{attachmentFileNames}{A \code{list} or \code{vector} of paths refering to the
attachements included in the email}
}
\description{
Build an email message from a vector or list of elements and
include any attachments.
}
\examples{
x = c(toHTML("Test", "h2"), br(), toHTML("test text"))
buildhtmlmsg(x)

}

