% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{buildModel}
\alias{buildModel}
\title{Easy to use version of xgboost}
\usage{
buildModel(data, response, modelType = "auto",
  calculationBudget = "balanced", trainProportion = 0.8,
  evalProportion = 0.2, onlyDataInd = FALSE)
}
\description{
Assumes default parameters for data configuration. Designed to hide complexity.
}

