% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionEvaluation.R
\name{check.glm.assumptions}
\alias{check.glm.assumptions}
\title{GLM Assumptions}
\usage{
check.glm.assumptions(model_obj, modelConfig = NULL)
}
\arguments{
\item{model_obj}{model object of the LM or GLM model under scrutiny}
}
\value{
The function returns a list. The list entries contain visuals and values illustrating the model's compliance with its relevant
model assumptions
}
\description{
This function checks the most prevalent model assumptions for LM and GLM models, including multi-collinearity, linearity,
distribution check (QQ-plot) and homoscedasticity
}
\examples{

fit <- glm(mpg ~ drat + wt, data=mtcars, family = gaussian())
assessment <- check.glm.assumptions(fit)
print(assessment)
assessment$chk_residuals
assessment$chk_homoscedasticity

}
\seealso{
\code{\link{glm}}, \code{\link{lm}}, \code{\link{pairs.panels}}
}

