% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{convertyDataToXgBoost}
\alias{convertyDataToXgBoost}
\title{Convert a training dataset (including target variable) into a list
of xgb.DMatrix objects.}
\usage{
convertyDataToXgBoost(x, target)
}
\arguments{
\item{x}{A list containing an element named 'trainData'. \code{x$trainData}
must contain the target variable}

\item{target}{A character string, the name of the target variable column}
}
\description{
Converts training dataset into xgboost datasets by
using \code{\link[xgboost]{xgb.DMatrix}}.
}
\examples{
DF1 <- data.frame(y = 1:6, a = letters[1:2], b = gl(3,2,labels = c("H","M","L")))
train.list <- list(trainData = DF1)
convertyDataToXgBoost(train.list, "y")
}

