% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{createGitIgnore}
\alias{createGitIgnore}
\title{Create and print out .gitignore}
\usage{
createGitIgnore(createFile = FALSE, openFile = TRUE, ignoreList = NULL)
}
\arguments{
\item{createFile}{create or overwrite the .gitignore file in current work directory}

\item{openFile}{open the .gitigore file or not}

\item{additional}{file types to be ignored by git}
}
\description{
A function to print out all suggested file types that need to be added to .gitignore file.
Or users can use the function to create or overwrite the .gitignore file directly under current work directory.
}
\examples{
createGitIgnore(createFile = TRUE, openFile = TRUE, ignoreList = c("*.RData", "*.jpg"))

}

