% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUtilities.R
\name{deployShinyApp}
\alias{deployShinyApp}
\title{Deploy a shiny app for all users}
\usage{
deployShinyApp(appDevPath, appTmpPath, appProdPath, users, appName, versionNo,
  changes, sendEmail = F)
}
\arguments{
\item{appDevPath}{The file path where the current dev version shiny app located}

\item{appTmpPath}{A temporary location that the shiny app will be moved to}

\item{appProdPath}{The production path that the shiny app will be deployed for the user}

\item{users}{A list of users of the shiny app}

\item{appName}{App's name}

\item{versionNo}{Version No of shiny app}

\item{changes}{Changes that have been updated for the shiny app}

\item{sendEmail}{Send an email to each user after the deployment completion}
}
\description{
Deploy a shiny app for all users
}

