% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalyDetection.R
\name{detect_anoms}
\alias{detect_anoms}
\title{Detects anomalies in a time series using S-H-ESD.}
\usage{
detect_anoms(data, k = 0.49, alpha = 0.05, num_obs_per_period = NULL,
  use_decomp = TRUE, use_esd = FALSE, one_tail = TRUE,
  upper_tail = TRUE, verbose = FALSE)
}
\arguments{
\item{data}{Time series to perform anomaly detection on.}

\item{k}{Maximum number of anomalies that S-H-ESD will detect as a percentage of the data.}

\item{alpha}{The level of statistical significance with which to accept or reject anomalies.}

\item{num_obs_per_period}{Defines the number of observations in a single period, and used during seasonal decomposition.}

\item{use_decomp}{Use seasonal decomposition during anomaly detection.}

\item{use_esd}{Uses regular ESD instead of hybrid-ESD. Note hybrid-ESD is more statistically robust.}

\item{one_tail}{If TRUE only positive or negative going anomalies are detected depending on if upper_tail is TRUE or FALSE.}

\item{upper_tail}{If TRUE and one_tail is also TRUE, detect only positive going (right-tailed) anomalies. If FALSE and one_tail is TRUE, only detect negative (left-tailed) anomalies.}

\item{verbose}{Additionally printing for debugging.}
}
\value{
A list containing the anomalies (anoms) and decomposition components (stl).
}
\description{
Detects anomalies in a time series using S-H-ESD.
}

