% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{emptyDF}
\alias{emptyDF}
\title{Create an empty data.frame}
\usage{
emptyDF(cols, N = 0)
}
\arguments{
\item{cols}{The column names of the data frame}

\item{N}{The number of rows to populate with NA}
}
\value{
A \code{data.frame} of dimension \code{N*length(cols)}
}
\description{
Create an empty data.frame
}
\examples{
emptyDF(c("A", "B"))
}

