% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{firstTimeSetup}
\alias{firstTimeSetup}
\title{Setup the RStuido Environment for New Users}
\usage{
firstTimeSetup()
}
\description{
Automatically generate \code{.Rprofile} and \code{.Renviron} for basic settings of RStudio development environment. The files will be created under user's root path (e.g. \code{'/home/user1/'})
For the first time user, please also see \code{'installCommonPackages()'} function for installing common packages.
}
\examples{
firstTimeSetup()

}
\seealso{
\link{installCommonPackages}
}

