% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptUtilities.R
\name{generateFeatureEffectsSlides}
\alias{generateFeatureEffectsSlides}
\title{Variable effects to ppt}
\usage{
generateFeatureEffectsSlides(doc, eff, nToShow = length(eff), rows = 1,
  cols = 1, baseTitle = "Feature effects")
}
\description{
uses \code{ReporteRs} to create slides
}
\examples{
df<-mtcars
df$am<-as.factor(df$am)
model<-glm(vs ~ am*mpg, data=df, family=binomial)
probs<-model$fit
eff = binaryClassifierEffect(probs,df,c('am','mpg'))

doc <-getAApowerpointTemplate()
doc = generateFeatureEffectsSlides(doc,eff, nToShow = 12, rows = 2, cols = 3)
writeDoc(doc, file = "test.pptx")
msg = buildhtmlmsg("", "test.pptx")
RQuantSendMail("ivan.liuyanfeng@gmail.com", subject = "test pptx", msg)
}

