% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyEmailReport.R
\name{generateForexSentimentReport}
\alias{generateForexSentimentReport}
\title{Weekly Forex Sentiment Email Report}
\usage{
generateForexSentimentReport(inst1, inst2, font_size = 6, image_size = 800,
  toAddress = c("ivan@xxx.com"))
}
\arguments{
\item{inst1}{currency symbol 1}

\item{inst2}{currency symbol 2}

\item{font_size}{size of font}

\item{image_size}{image size embeded in email}

\item{toAddress}{a vector of email addresses}
}
\description{
Weekly Forex Sentiment Email Report
}
\examples{
generateForexSentimentReport(inst1 = "AUD", inst2 = "USD",
font_size = 10, image_size = 700,
toAddress = c("ivan@xxx.com"))

}

