% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyEmailReport.R
\name{generateOandaCointegrationReport}
\alias{generateOandaCointegrationReport}
\title{Oanda Weekly Cointegrated Pairs Searc}
\usage{
generateOandaCointegrationReport(total_points = 250, testPeriod = 0.05,
  toAddress = c("ivan@xxx.com"))
}
\arguments{
\item{total_points}{total data points (days) to look back for testing}

\item{testPeriod}{0-1, proportion of data points to be used as testing set}

\item{toAddress}{a vector of email addresses}
}
\description{
Weekly Search Cointegrated Pairs  over all tradable Forex pairs on Oanda Platform
}
\examples{
generateOandaCointegrationReport(total_points = 250, testPeriod = 0.05, toAddress = c("ivan@xxx.com"))

}

