% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClassification.R
\name{generatePrecisionRecallCurve}
\alias{generatePrecisionRecallCurve}
\title{Generate a standard Precision/Recall plot}
\usage{
generatePrecisionRecallCurve(pred, header = "")
}
\arguments{
\item{pred}{A \code{prediction} object.}

\item{header}{An optional string to add to the plot title}
}
\value{
A ggplot
}
\description{
Evaluate a model with a Precision vs Recall plot
}
\examples{
set.seed(1234)
probs <- pnorm(rnorm(100,-1,1))
actuals <- rbinom(100,1,probs)
pred <- prediction(probs, actuals)
generatePrecisionRecallCurve(pred, header = "A Model")
}
\seealso{
\code{\link{binaryClassifierEvaluation}}
}

