% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClassification.R
\name{generateROCCurve}
\alias{generateROCCurve}
\title{Generate a standard ROC curve}
\usage{
generateROCCurve(pred, header = "")
}
\arguments{
\item{pred}{A \code{prediction} object.}

\item{header}{An optional string to add to the plot title}
}
\value{
A ggplot
}
\description{
Evaluate a model with a ROC curve
}
\examples{
set.seed(1234)
probs <- pnorm(rnorm(100,-1,1))
actuals <- rbinom(100,1,probs)
pred <- prediction(probs, actuals)
generateROCCurve(pred, header = "A Model")

}
\seealso{
\code{\link{binaryClassifierEvaluation}}
}

