% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptUtilities.R
\name{getRQuantpowerpointTemplate}
\alias{getRQuantpowerpointTemplate}
\title{The RQuant standard powerpoint template}
\usage{
getAApowerpointTemplate()
}
\description{
Read in the RQuant powerpoint template
}

