% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{getDataFrameVariableTypes}
\alias{getDataFrameVariableTypes}
\title{Extract variable types from data.frames}
\usage{
getDataFrameVariableTypes(x)
}
\arguments{
\item{x}{A \code{data.frame}}
}
\description{
This function extracts the variable classes from a \code{data.frame}.
}
\examples{
getDataFrameVariableTypes(mtcars)

}

