% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureEngineering.R
\name{getFinancialWeekFromDate}
\alias{getFinancialWeekFromDate}
\title{Get the Financial Week from a particular Date
(Not working yet!)
This is used for the extractDateFeatures default function for retrieving the financial year. Based on the financial year starting on 1st July of every year.}
\usage{
getFinancialWeekFromDate(date)
}
\arguments{
\item{date}{A date containing a vector of dates to compute the financial year for.}
}
\value{
The financial year associated to each date.
}
\description{
Get the Financial Week from a particular Date
(Not working yet!)
This is used for the extractDateFeatures default function for retrieving the financial year. Based on the financial year starting on 1st July of every year.
}
\examples{
getFinancialWeekFromDate(as.Date("01/07/2016","\%d/\%m/\%Y"))
getFinancialWeekFromDate(as.Date("26/06/2016","\%d/\%m/\%Y"))
}

