% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureEngineering.R
\name{getFinancialYearFromDate}
\alias{getFinancialYearFromDate}
\title{Get the Financial Year from a particular Date}
\usage{
getFinancialYearFromDate(date)
}
\arguments{
\item{date}{A date containing a vector of dates to compute the financial year for.}
}
\value{
The financial year associated to each date.
}
\description{
This is used for the extractDateFeatures default function for retrieving the financial year. Based on the financial year starting on 1st July of every year.
}
\examples{
getFinancialYearFromDate(as.Date("26/06/2016","\%d/\%m/\%Y"))
getFinancialYearFromDate(as.Date("26/07/2016","\%d/\%m/\%Y"))
}

