% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtilities.R
\name{getJSONFile}
\alias{getJSONFile}
\title{Get a JSON file}
\usage{
getJSONFile(path)
}
\arguments{
\item{path}{A path to pass to \code{\link{file}} which is then passed to
\code{\link{readlines}}}
}
\description{
Get a JSON file
}

