% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureEngineering.R
\name{getSeasonFromDate}
\alias{getSeasonFromDate}
\title{Get the season for a particular Date}
\usage{
getSeasonFromDate(date)
}
\arguments{
\item{date}{A date containing a vector of dates to compute the seasons for.}
}
\value{
The season (Summer, Autumn, Winter, Spring) as a character vector.
}
\description{
This is used for the extractDateFeatures default function for retrieving a season. This is currently using Australian-based seasons.
}
\details{
It will assume the following mapping:
 Dec - Feb: Summer
 Mar - May: Autumn
 Jun - Aug: Winter
 Sep - Nov: Spring
}
\examples{
getSeasonFromDate(Sys.Date())
}

