% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{getTrainTestIds}
\alias{getTrainTestIds}
\title{function to take in a set of ids and produce a list of training
and test IDs using random sampling}
\usage{
getTrainTestIds(x, trainProportion, evalProportion = NULL, seed = 999,
  verbose = F)
}
\description{
function to take in a set of ids and produce a list of training
and test IDs using random sampling
}

