% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEmail.R
\name{htmlTableQuant}
\alias{htmlTableQuant}
\title{Create HTML table}
\usage{
htmlTableQuant(dat, showRowNames = FALSE, cornerText = "")
}
\arguments{
\item{showRowNames}{A boolean flag on whether to show the row names from data frame}

\item{cornerText}{A string to display in the top left cell when \code{showRowNames} is used}

\item{x}{A \code{data.frame}}
}
\description{
This function creates a HTML table from a data frame and applies
RQuant branding standards (eg colours / fonts etc) so that it
can be fowared on to clients
}
\details{
Newlines have been added to prevent the output falling foul of
the 1000-character line limit of SMTP.  The mailserver inserts
its own newlines, breaking the code.
}
\examples{
htmlTableQuant(mtcars)
htmlTableQuant(mtcars, showRowNames = TRUE)
htmlTableQuant(mtcars, showRowNames = TRUE, cornerText = "X")

}

