% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{installCommonPackages}
\alias{installCommonPackages}
\title{Install Common Packages}
\usage{
installCommonPackages(update = FALSE)
}
\arguments{
\item{update}{if the library exists, try to update the library instead. Default \code{FALSE}}
}
\description{
Automatically check and install a list of common packages required for financial modeling, data science, statistical analysis and etc. And choose to update existing libraries.
Please make sure you've already run \code{'firstTimeSetup()'} before installing packages. The function helps you to re-locate your library path to a shared folder.
}
\examples{
Sys.getenv('R_LIBS_USER')
installCommonPackages(update = FALSE)

}
\seealso{
\link{firstTimeSetup}
}

