% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{listFns}
\alias{listFns}
\title{List functions within a loaded package}
\usage{
listFns(pkg, pattern = NULL, invert = FALSE)
}
\arguments{
\item{pkg}{The name of a package as a string}

\item{pattern}{A regex expression to pass to grep to
include only certain package}

\item{invert}{A logical indicating whether to return names that match the
provided pattern (TRUE) or that do not match the pattern (FALSE)}
}
\value{
A vector of function names within a loaded package
}
\description{
List functions within a loaded package
}
\examples{
library(ggplot2)

listFns("ggplot2")
listFns("ggplot2", pattern = "^theme")
listFns("ggplot2", pattern = "^[Stat|theme|geom|Geom|scale]", invert = TRUE)

}

