% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{mergeDataFrames}
\alias{mergeDataFrames}
\title{Merge a list of data.frames}
\usage{
mergeDataFrames(datasets, mandatoryMergeFields, optionalMergeFields = NULL)
}
\arguments{
\item{datasets}{a \code{list} of data.frames with column names}

\item{mandatoryMergeFields}{The fields that will be used in the join}

\item{optionalMergeFields}{The fields that will be used in the join if they exist}
}
\description{
This function takes in a list of data sets and merges them all together as a
series of left joins. There are some mandatory merge fields and some optional
fields that we will try to merge on if they exist. We assume that the first
element in the list is complete and has both the optional and the mandatory
columns.
}
\examples{
d1 = mtcars
d2 = mtcars
colnames(d2)[!colnames(d2)\%in\%c("gear","carb")] = LETTERS[1:(ncol(d2)-2)]
View(mergeDataFrames(list(d1, d2), c("gear", "carb")))
}

