% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textMining.R
\name{ngramify}
\alias{ngramify}
\title{Generate n-grams from a document}
\usage{
ngramify(ngram_df, split_num = 1, grams = 3)
}
\arguments{
\item{ngram_df}{a data.frame with strings to be tokenized.}

\item{split_num}{number of chunks}

\item{grams}{n grams}
}
\description{
splits strings into n-grams with given minimal and maximal numbers of grams.
}
\examples{
setupTwitterConn()
tweets <- tweet_corpus(search = "audusd", n = 100, since = as.character(Sys.Date()-7), until = as.character(Sys.Date()))
tweets_ngrams <- ngramify(tweets$d, split_num=1, grams = 3)

}
\seealso{
\link{tfidfTransformer}
}

