% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{optimiseHyperparameters}
\alias{optimiseHyperparameters}
\title{Based on the code from rBayesianOptimisation}
\usage{
optimiseHyperparameters(FUN, data, modelparams, bounds, init_points, n_iter,
  acq = "ucb", kappa = 2.576, eps = 0, verbose = TRUE, ...)
}
\description{
Based on the code from rBayesianOptimisation
}

