% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTimings.R
\name{prepareTimingTable}
\alias{prepareTimingTable}
\title{Convert a timings table to a final timings table}
\usage{
prepareTimingTable(timingsTable)
}
\arguments{
\item{timingsTable}{A timingsTable created by \code{addToTimingsTable}}
}
\value{
A data frame containing the tasks performed, when they were
started and how long they took to run
}
\description{
Take in a table created by \code{addToTimingsTable} and add useful information to it like time elapsed
}
\examples{
timingsTable = addToTimingsTable("Start")
timingsTable = addToTimingsTable("Step 1", timingsTable)
Sys.sleep(4)
prepareTimingTable(timingsTable)

}

