% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitVersioning.R
\name{pushGitRepo}
\alias{pushGitRepo}
\title{Add Commit and Push Cuurent Git Repository}
\usage{
pushGitRepo(gitURL = "git@github.com:ivanliu1989/RQuant.git",
  initRepo = FALSE, branch = "master", username = "ivanliu1989",
  commitMsg = "update repo", repoPath = getwd())
}
\arguments{
\item{gitURL}{the SSH URL of git repository}

\item{initRepo}{initialise the git repo or not}

\item{branch}{default "master"}

\item{username}{username of git account}

\item{commitMsg}{commit message}

\item{repoPath}{local path of repository}
}
\description{
Add Commit and Push Cuurent Git Repository
}
\examples{
pushGitRepo(gitURL = "git@github.com:ivanliu1989/RQuant.git",
branch = "master",
commitMsg = "update repo",
repoPath = getwd())

}

