% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtilities.R
\name{readSQLFileAndUpdateParameters}
\alias{readSQLFileAndUpdateParameters}
\title{Read in a sql file and update the parameters}
\usage{
readSQLFileAndUpdateParameters(path, dataConfig = list())
}
\arguments{
\item{path}{A file path that refers to a sql file}

\item{dataConfig}{The configuration list, where the names represent the
parameter and the values are what will be replaced. will be passed to \code{updateSQLwithParams}}
}
\description{
Given a file path to a sql query, read in the query and update based
on a set of parmeters
}

