% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionEvaluation.R
\name{regressionEvaluation}
\alias{regressionEvaluation}
\title{Evaluate Regression Model Results}
\usage{
regressionEvaluation(obs, pred, model = NULL, sample_size = 1000,
  seed = 1234)
}
\arguments{
\item{obs}{A vector of actual outcomes.}

\item{pred}{A vector of fitted values.}

\item{model}{Optional, the model used to predict \code{fitted values} from \code{actual outcomes}.}

\item{sample_size}{the maximum sample size, in numbers or percents of total observations, will be used to visualise regression effects.}

\item{seed}{Random seeds for reproducibility.}
}
\value{
A \code{list} of \code{data.table} and \code{ggplot2}
}
\description{
Evaluate the performance of a regression model.
}
\examples{
data(mtcars)
fit <- glm(mpg ~ cyl + disp + hp + drat, data=as.data.frame(mtcars), family = gaussian())
actuals <- mtcars$mpg
probs <- predict(fit, mtcars)
res <- regressionEvaluation(pred = probs,
obs = actuals,
model = fit,
sample_size=1e5,
seed = 1234)
res$descriptive_statistics
res$goodness_of_fit
res$model_details
res$visualisation$residual_vs_fitted
res$visualisation$residual_vs_order
res$visualisation$histograms
res$visualisation$prediction_intervals
res$check_model_assumptions

}

