% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textMining.R
\name{replacePunctuation}
\alias{replacePunctuation}
\title{Replace Punctuation}
\usage{
replacePunctuation(strings)
}
\arguments{
\item{strings}{strings}
}
\description{
Replace punctuation with a " " in charactors
}
\seealso{
\link{tfidfTransformer}
}

