% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{roughModelDataSetToNumeric}
\alias{roughModelDataSetToNumeric}
\title{Rough conversion of data set to numeric data set}
\usage{
roughModelDataSetToNumeric(modellingData)
}
\description{
Rough conversion of data set to numeric data set
}

