% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEmail.R
\name{RQuantSendMail}
\alias{RQuantSendMail}
\title{Send an Email through R}
\usage{
RQuantSendMail(to = c("ivan@xxx.com"), subject = NULL,
  msg = NULL, attach.files = NULL, file.names = NULL,
  file.descriptions = NULL)
}
\arguments{
\item{to}{A character vector of recipient valid email addresses.}

\item{subject}{Subject of the email.}

\item{msg}{Body of the email as text. If the parameter body refers to an existing file location, the text of the file is parsed as body of the email.}

\item{attach.files}{A character vector of paths in the file system linking to files or *valid* URLs to be attached to the email (see details for more info on attaching URLs)}

\item{file.names}{A character vector of names of attached files}

\item{file.descriptions}{A character vector of descriptions of attached files}
}
\description{
Send an email through R.
}
\examples{
RQuantSendMail(to = "ivan@xxx.com", subject = "An email from R", msg = "test", attach.files = NULL)

}

