% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{splitDataToTrainTestDataFrame}
\alias{splitDataToTrainTestDataFrame}
\title{Function to split a dataset and return training, test and eval sets as dataframes}
\usage{
splitDataToTrainTestDataFrame(x, trainProportion, evalProportion = NULL,
  seed = 999, verbose = F)
}
\description{
Function to split a dataset and return training, test and eval sets as dataframes
}

