% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{splitDataToTrainTestDataFrameFromId}
\alias{splitDataToTrainTestDataFrameFromId}
\title{Function to split a dataset and return training, test and eval sets as dataframes}
\usage{
splitDataToTrainTestDataFrameFromId(x, IdVar, trainProportion,
  evalProportion = NULL, seed = 999, verbose = F)
}
\description{
Function to split a dataset and return training, test and eval sets as dataframes
}

