% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{splitDataToTrainTestStratified}
\alias{splitDataToTrainTestStratified}
\title{Function to do stratified sampling for model data}
\usage{
splitDataToTrainTestStratified(x, strataCols = NULL, trainProportion,
  evalProportion = NULL, levels = 50, seed = 999, DT = T)
}
\arguments{
\item{x}{A data.frame or data.table object containing all data to be split}

\item{strataCols}{Names of columns or features that will be used to do stratification}

\item{trainProportion}{Proportion of training set}

\item{evalProportion}{Proportion of validation set}

\item{levels}{Threshold of determine a feature to be categorical and qualified for stratification}

\item{seed}{Random seeds for reproducibility}

\item{DT}{If the function returns a list of data.tables or data.frames}
}
\value{
A \code{list} of \code{data.table} or \code{data.frame}
}
\description{
Function to stratified split a dataset and return training, test and eval sets as data.frame or data.table
}
\examples{
data(mtcars)
res <- splitDataToTrainTestStratified(mtcars, names(mtcars), 0.7, 0.3, 15, 999, T)
res$trainData
res$testData
res$evalData

}

