% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{summarise.data.frame}
\alias{summarise.data.frame}
\title{Summary Dataframe}
\usage{
summarise.data.frame(df)
}
\arguments{
\item{df}{The data frame to be summarised}
}
\value{
The function returns a list. The list entries contain tabel-level and column-level results
}
\description{
This function takes a data frame as an input, and provides a quick summary of the contents of the data frame.
It provides table-level (size and last date updated) and column-level (e.g. minimum, maximum, missing values) information
For character fields or factors, it provides the number of distinct values. For continuous values, it shows the minimum and maximum.
In addition it returns an example value, the number of missing values, the technical and statitical data type
}
\examples{
res <- summarise.data.frame(mtcars)
res$table_level
res$column_level
print(res)

}
\seealso{
\code{\link{str}}, \code{\link{head}} \code{\link{summary.data.frame}}
}

