% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulations.R
\name{summarise.df}
\alias{summarise.df}
\title{Summary Dataframe}
\usage{
summarise.df(df, filename = "output", round.date = TRUE, n.values = 10,
  n.print = 8, n.pairs = 12)
}
\arguments{
\item{df}{The data frame to be summarised}

\item{filename}{A character string to use as the name of the output files}

\item{round.date}{A logical, whether to round datetime values to just date}

\item{n.values}{Integer, the maximum number of distinct values to show for frequency tables.  For fields with more than
this number of distinct values, a frequency table will be displayed.  For fields with less than this number of
distinct values, the first few values will be displayed.}

\item{n.print}{Integer, the number of sample values to display.}

\item{n.pairs}{Integer, the number of numeric variables to display in the pairs plot.}
}
\value{
The function does not return any value.  But it does produce a text file and a png file, based on the supplied
 filename stub.
}
\description{
This function takes a data frame as an input, and provides a quick summary of the contents of the data frame.

For character fields or factors, it provides a frequency table, sorted decreasing, if there are relatively few
distinct values.  Otherwise, the first few values are shown.

For numeric fields, including date or datetime fields, it provides quartiles.

It also provides a pairs plots for the first few numeric fields.
}
\examples{
summarise.df(mtcars)

}
\seealso{
\code{\link{str}}, \code{\link{summary.data.frame}}, \code{\link{head}}
}

