% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{theme_rquant}
\alias{theme_rquant}
\title{Use a ggplot2 theme with r aesthetics}
\usage{
theme_rquant(simple_theme = TRUE, r_font = TRUE, r_colours = TRUE,
  font_size = 12L, logo = "none", opacity = 0.1, position = 0L,
  inv = FALSE)
}
\arguments{
\item{simple_theme}{logical. Use a simpler theme? default is TRUE.}

\item{r_font}{logical. Use the r font? (Bold Calibri is used as a stand-in)}

\item{r_colours}{logical. Use the r colours for colour and fill?}

\item{font_size}{text size (in pts)}

\item{logo}{character. one of \code{c('none','r')}}

\item{opacity}{numeric. 0 to 1, default 0.5}

\item{position}{an integer usually less than 8. different logos have different options. The first 5 options are: \cr
0 - centre, large \cr
1 - top-left, medium \cr
2 - top-right, medium \cr
3 - bottom-right, medium \cr
4 - bottom-left, medium}

\item{inv}{1,2,3 color themes}
}
\description{
Adds a nice background, changes colours (and fill colours) to utilise
r colours, changes text to a r font standin, changes text size,
and adds an optional r logo. The logo can be rendered at half
or full opacity and can be renderd in the centre or the corner at different sizes. \cr
If you want to render the logo \emph{behind} the graphical objects just put your
 call to \code{\link{theme_rquant}} earlier than your call to those objects.
}
\examples{
\dontrun{
set.seed(1)
dt <- data.frame(x = 1:10, y = 1:10, z = runif(10)*10, c  = sample(LETTERS[1:5],10, replace = TRUE))
dt

p <- ggplot(dt) +
geom_bar(aes(x= x, y = y, fill = rev(c)), stat = "identity") +
geom_point(aes(x= x, y = z, col = c), size = 5) +
ggtitle("r Plot")

p

# Add r colours
p + theme_rquant(simple_theme = FALSE, r_font = FALSE)

# Add a visually pleasing theme (theme_minimal())
p + theme_rquant(r_font = FALSE)

# Add r font
p + theme_rquant()

# Add the r logo
p + theme_rquant(logo = "r")

# Add a transparent logo
p + theme_rquant(logo = "r", opacity = 0.3)

# change the logos position and size
p + theme_rquant(logo = "r", opacity = 1, position = 1)
p + theme_rquant(logo = "r", opacity = 1, position = 3)

# Add a logo which sits 'behind' the graphical objects
ggplot(dt) +
theme_rquant(logo = "r", opacity = 0.5, position = 0) +
geom_bar(aes(x= x, y = y, fill = rev(c)), stat = "identity") +
geom_point(aes(x= x, y = z, col = c), size = 5) +
ggtitle("r Plot")

}

}

