% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEmail.R
\name{toHTML}
\alias{toHTML}
\title{Convert text to HTML element}
\usage{
toHTML(x, tag = "p")
}
\arguments{
\item{x}{A string of text}

\item{tag}{Defaults to "p". May be any of "a", "em", "strong", "cite", "q", "dfn", "abbr", "data", "time", "code", "var", "samp", "kbd", "mark", "ruby", "rb", "rt", "rp", "rtc", "bdi", "bdo", "span", "br", "wbr", "small", "i", "b", "u", "s", "sub", "sup", "p", "pre", "blockquote", "ol", "ul", "li", "dl", "dt", "dd", "figure", "figcaption", "div", "main", "hr"}
}
\description{
Convert text to HTML element
}
\examples{
toHTML("hello")
toHTML("hello", "h2")

}

