% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textMining.R
\name{topicModeling}
\alias{topicModeling}
\title{Topic Modeling}
\usage{
topicModeling(corpus, k = 10)
}
\arguments{
\item{corpus}{the words}

\item{k}{number of topics}
}
\description{
Topic Modeling based on Correlated Topic Model (Estimate a CTM model using for example the VEM algorithm.) and
Latent Dirichlet Allocation (Estimate a LDA model using for example the VEM algorithm or Gibbs Sampling.)
}
\details{
LDA:\cr
The C code for LDA from David M. Blei and co-authors is used to estimate and fit a latent dirichlet allocation model with the VEM algorithm. For Gibbs Sampling the C++ code from Xuan-Hieu Phan and co-authors is used.\cr
When Gibbs sampling is used for fitting the model, seed words with their additional weights for the prior parameters can be specified in order to be able to fit seeded topic models.\cr
\cr
CTM:\cr
The C code for CTM from David M. Blei and co-authors is used to estimate and fit a correlated topic model.\cr
}
\examples{
setupTwitterConn()
tweets <- tweet_corpus(search = "audusd", n = 100, since = as.character(Sys.Date()-7), until = as.character(Sys.Date()))
tweets <- text_clean(tweets$v, rmDuplicates = FALSE, cores = 6, stems = NULL)
wordCloudVis(tweets$corpus)
topicModeling(tweets$corpus, k = 5)

}

