% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtilities.R
\name{updateSQLwithParams}
\alias{updateSQLwithParams}
\title{Parameterised SQL.}
\usage{
updateSQLwithParams(x, dataConfig)
}
\arguments{
\item{x}{A string that represents a query with @paramName to be replaced.}

\item{dataConfig}{The configuration list, where the names represent the
parameter and the values are what will be replaced.}
}
\value{
A string that represents an SQL query.
}
\description{
The core parameter replacement function accepting a list
of parameter names to replace and replacing all the
occurances in x that have @......
}
\note{
Dates have to use the format that is expected by SQL.
}
\examples{
updateSQLwithParams("select * from @table", list(table = "blah"))

query.txt<-"
select
   sum(total_sales_amount) as spend
   ,transaction_date
from
   cedwdm.transaction_product_vw
where
   transaction_date > '@start.date'
group by
   transaction_date
"
my.query<-updateSQLwithParams(
   query.txt
   , list(start.date = format(Sys.Date() - 7, "\%d-\%b-\%Y" ))
   )
print(my.query)

}
\seealso{
\code{\link{format}}, \code{\link{dbGetQuery}}
}

