% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUtilities.R
\name{userDeploymentEmail}
\alias{userDeploymentEmail}
\title{Send a standard release email to a user of a shiny app}
\usage{
userDeploymentEmail(user, changes, appName, versionNo, appProdPath)
}
\arguments{
\item{user}{User details of the shiny app}

\item{changes}{Changes that have been updated for the shiny app}

\item{appName}{App's name}

\item{versionNo}{Version No of shiny app}

\item{appProdPath}{The production path that the shiny app will be deployed for the user}
}
\description{
Send a standard release email to a user of a shiny app
}

