% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textMining.R
\name{wordCloudVis}
\alias{wordCloudVis}
\title{Plot a word cloud}
\usage{
wordCloudVis(corpus, min.freq = 2, max.words = 150, random.order = TRUE)
}
\arguments{
\item{corpus}{the words}

\item{min.freq}{words with frequency below min.freq will not be plotted}

\item{max.words}{Maximum number of words to be plotted. least frequent terms dropped}

\item{random.order}{plot words in random order. If false, they will be plotted in decreasing frequency}
}
\description{
Plot a word cloud
}
\details{
If freq is missing, then words can either be a character vector, or Corpus. If it is a vector and freq is missing, standard stop words will be removed prior to plotting.
}
\examples{
setupTwitterConn()
tweets <- tweet_corpus(search = "audusd", n = 100, since = as.character(Sys.Date()-7), until = as.character(Sys.Date()))
tweets <- text_clean(tweets$v, rmDuplicates = FALSE, cores = 6, stems = NULL)
wordCloudVis(tweets$corpus)

}
\seealso{
\link{tweet_corpus}

\link{tweet_clean}

\link{tweet_clean_corpus}
}

