% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{xgboostModel}
\alias{xgboostModel}
\title{Run an eXtreme Gradient Boosting model}
\usage{
xgboostModel(datasets, target, paramxgBoost = NULL, nrounds = 100,
  early.stop.round = 5, print.every.n = 50, verbose = 1)
}
\arguments{
\item{datasets}{a list of train, test and validation datasets}

\item{target}{the name of the target feature in datasets}

\item{paramxgBoost}{a list of parameters for xgboost}

\item{nrounds}{the max number of iterations}

\item{early.stop.round}{If NULL, the early stopping function is not triggered. If set to an integer k, training with a validation set will stop if the performance keeps getting worse consecutively for k rounds.}

\item{print.every.n}{Print every N progress messages when verbose>0. Default is 1 which means all messages are printed.}

\item{verbose}{If 0, xgboost will stay silent. If 1, xgboost will print information of performance. If 2, xgboost will print information of both}
}
\description{
Run an eXtreme Gradient Boosting model
}
\examples{
set.seed(42)
x = iris
x$Species = ifelse(x$Species == "versicolor",1, 0)
dat = splitDataToTrainTestDataFrame(x,.9,.25)
mod = xgboostModel(dat,"Species",nrounds = 200, early.stop.round = 5, verbose = 1)

}

