% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{xgboostModel_alpha}
\alias{xgboostModel_alpha}
\title{Train an eXtreme Gradient Boosting model [Alpha]}
\usage{
xgboostModel_alpha(data, response, modelType = "r", dataConfig = list(),
  paramConfig = list(), nrounds = 250, earlyStopRatio = 0.25,
  objective = NULL, eval_metric = NULL, onlyDataInd = FALSE,
  opt.bayesian = TRUE, opt.iter = 10, opt.initialpts = 10,
  opt.savemodel = TRUE, buildEnsemble = TRUE)
}
\arguments{
\item{data}{The dataset to train the xgboost model on. Should be a list containing training, validation and test sets (if applicable).}

\item{response}{The type of xgboost model to train - currently supports either 'r' (regression) or 'c' (classification).}

\item{dataConfig}{A list containing the data preprocessing configuration settings. Refer to details for more information.}

\item{paramConfig}{A list containing the bounds of the xgboost model to optimise the parameters for.}

\item{nrounds}{The number of rounds to run xgboost for. This should be set to a higher number if eta is fairly low.}

\item{earlyStopRatio}{The early stopping condition, where xgboost will stop training if it fails to find a better validation score for x of number of rounds.}

\item{objective}{The objective for the xgboost model. Leave NULL for automatic determination. Default: NULL.}

\item{eval_metric}{The evaluation metric for the xgboost model. Leave NULL for automatic determination. Default: NULL.}

\item{onlyDataInd}{A logical containing whether or not to build the model or stop after data preprocessing. Can be used to derive the test dataset using the same transformation as the training dataset. Default: FALSE.}

\item{opt.iter}{The number of Bayesian optimisation iterations to perform. Default: 10.}

\item{opt.initialpts}{The initial number of points to initialise the Bayesian optimisation for. Default: 10.}

\item{opt.savemodel}{A logical containing whether or not to save the models being trialed in the Bayesian optimisation process. Default: TRUE.}

\item{train.proportion}{The splitting ratio for train/validation. Note that this is not stratified (future enhancement).}
}
\value{
A list containing two elements. Models which contain all the xgb boost models built as part of the hyperparameter optimisation. Results which contain the Bayesian optimisation output.
}
\description{
Train an eXtreme Gradient Boosting model [Alpha]
}

