% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{xgboostVariableImportance}
\alias{xgboostVariableImportance}
\title{Get the variable Importance of a xgboost Model}
\usage{
xgboostVariableImportance(mod, N = 30, allowedVariables = NULL)
}
\arguments{
\item{mod}{a Model from \code{xgboostModel}}

\item{N}{The Number of features to return}

\item{allowedVariables}{Features that are allowed to be included}
}
\value{
A list containing a data.table with the variable importance scores and a ggplot
}
\description{
Generic Function to get the importance of features in a xgboost model
}
\examples{
set.seed(42)
x = iris
x$Species = ifelse(x$Species == "versicolor",1, 0)
dat = splitDataToTrainTestDataFrame(x,.9,.25)
mod = xgboostModel(dat,"Species")
v = xgboostVariableImportance(mod)

v = xgboostVariableImportance(mod, 30, allowedVariables = c("Sepal.Length","Sepal.Width"))

}

